r.flow: rename io.h file which causes conflict with MSYS2
Closes: https://github.com/OSGeo/grass/issues/3700
